Moving back end tests into a server folder
