+++ test/server/models/user.js
@@ -1,9 +1,9 @@
 var chai = require('chai');
-var db = require('../../server/lib/db');
+var db = require('../../../server/lib/db');
 chai.should();
 
 describe('user', function() {
-  var user = require('../../server/models/user');
+  var user = require('../../../server/models/user');
 
   var githubProfile = {
     _json: {
